+++ src/main/java/webserver/http/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
