+++ src/main/java/webserver/resolver/ResourceResolver.java
@@ -2,8 +2,8 @@ package webserver.resolver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.http.HttpRequest;
-import webserver.http.HttpResponse;
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
 import webserver.http.common.ContentType;
 
 import java.io.IOException;
