+++ src/test/java/webserver/http/HttpRequestTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webserver.http.exception.RequestParseException;
+import webserver.http.request.HttpRequest;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
