+++ src/test/java/webserver/http/HttpResponseTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.http.response.HttpResponse;
 
 import java.io.ByteArrayOutputStream;
 import java.io.DataOutputStream;
