+++ src/test/java/webserver/resolver/ResourceResolverTest.java
@@ -1,7 +1,7 @@
 package webserver.resolver;
 
-import webserver.http.HttpRequest;
-import webserver.http.HttpResponse;
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
